+++ crates/core/flags/complete/mod.rs
@@ -1,0 +1,8 @@
+/*!
+Modules for generating completions for various shells.
+*/
+
+pub(super) mod bash;
+pub(super) mod fish;
+pub(super) mod powershell;
+pub(super) mod zsh;