+++ crates/core/logger.rs
@@ -1,7 +1,10 @@
-// This module defines a super simple logger that works with the `log` crate.
-// We don't need anything fancy; just basic log levels and the ability to
-// print to stderr. We therefore avoid bringing in extra dependencies just
-// for this functionality.
+/*!
+Defines a super simple logger that works with the `log` crate.
+
+We don't do anything fancy. We just need basic log levels and the ability to
+print to stderr. We therefore avoid bringing in extra dependencies just for
+this functionality.
+*/
 
 use log::{self, Log};
 
@@ -10,15 +13,16 @@ use log::{self, Log};
 /// This logger does no filtering. Instead, it relies on the `log` crates
 /// filtering via its global max_level setting.
 #[derive(Debug)]
-pub struct Logger(());
+pub(crate) struct Logger(());
 
+/// A singleton used as the target for an implementation of the `Log` trait.
 const LOGGER: &'static Logger = &Logger(());
 
 impl Logger {
     /// Create a new logger that logs to stderr and initialize it as the
     /// global logger. If there was a problem setting the logger, then an
     /// error is returned.
-    pub fn init() -> Result<(), log::SetLoggerError> {
+    pub(crate) fn init() -> Result<(), log::SetLoggerError> {
         log::set_logger(LOGGER)
     }
 }