+++ crates/searcher/src/searcher/mod.rs
@@ -124,7 +124,7 @@ impl BinaryDetection {
 /// source data from an encoding to UTF-8 before searching.
 ///
 /// An `Encoding` will always be cheap to clone.
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Encoding(&'static encoding_rs::Encoding);
 
 impl Encoding {