+++ tests/regression.rs
@@ -402,6 +402,7 @@ rgtest!(r428_unrecognized_style, |dir: Dir, mut cmd: TestCommand| {
     let output = cmd.cmd().output().unwrap();
     let stderr = String::from_utf8_lossy(&output.stderr);
     let expected = "\
+error parsing flag --colors: \
 unrecognized style attribute ''. Choose from: nobold, bold, nointense, \
 intense, nounderline, underline.
 ";